refactor: refactor helpers
