+++ src/data/usersData.json
@@ -1,1 +1,1 @@
-[{"id":1,"name":"Aleksandr","password":"salainen","wins":323},{"id":2,"name":"owner","password":"55555","wins":61},{"id":3,"name":"user@example.com","password":"qwerty123","wins":307},{"id":999,"name":"BOT","password":"bot","wins":0}]
+[{"id":1,"name":"Aleksandr","password":"salainen","wins":324},{"id":2,"name":"owner","password":"55555","wins":61},{"id":3,"name":"user@example.com","password":"qwerty123","wins":307},{"id":999,"name":"BOT","password":"bot","wins":0}]
