+++ src/web_socket_server/modules/singlePlay.ts
@@ -1,7 +1,6 @@
 import { CustomWebSocket, RoomType, ShipType } from '../../app/types.js';
-
 import { BOT_ID, FiELD_SIZE, SEA, SHIP } from '../../app/variables.js';
-import { createResponse } from '../../app/healpers.js';
+import { createResponse } from '../helpers/responses.js';
 import { rooms } from '../../data/rooms.js';
 import { users } from '../../data/users.js';
